backend/requirements.txt
# =============================================================================
# VeriField Nexus — Backend Dependencies
# =============================================================================

# --- Core Framework ---
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9

# --- Database ---
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.0

# --- Validation & Settings ---
pydantic==2.9.0
pydantic-settings==2.5.0

# --- Authentication ---
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# --- Supabase ---
supabase==2.9.0
httpx==0.27.0

# --- Image Processing ---
imagehash==4.3.1
Pillow==10.4.0

# --- AI / ML ---
scikit-learn>=1.5.0
numpy>=1.26.0

# --- Utilities ---
python-dotenv==1.0.1
